Reorganization of classes dealing with connecting to the old itc
