+++ modules/itc/src/main/scala/lucuma/itc/ItcImpl.scala
@@ -93,6 +93,9 @@ object ItcImpl {
         exposures:        Int,
         level:            NonNegInt
       ): F[ItcResult] =
+        import lucuma.itc.legacy.given
+        import lucuma.itc.legacy.*
+
         Trace[F].span("legacy-itc-query") {
           val json =
             spectroscopyParams(targetProfile,
@@ -135,6 +138,9 @@ object ItcImpl {
         exposureDuration: Quantity[BigDecimal, Second],
         exposures:        Int
       ): F[ItcGraphResult] =
+        import lucuma.itc.legacy.given
+        import lucuma.itc.legacy.*
+
         Trace[F].span("legacy-itc-query") {
           val json =
             spectroscopyParams(targetProfile,
@@ -293,32 +299,5 @@ object ItcImpl {
 
       }
     }
-
-  /** Convert model types into OCS2 ITC-compatible types for a spectroscopy request. */
-  private def spectroscopyParams(
-    targetProfile:    TargetProfile,
-    observingMode:    ObservingMode,
-    exposureDuration: FiniteDuration,
-    conditions:       ItcObservingConditions,
-    exposures:        Int
-  ): ItcParameters =
-    ItcParameters(
-      source = ItcSourceDefinition.fromTargetProfile(targetProfile),
-      observation = ItcObservationDetails(
-        calculationMethod = ItcObservationDetails.CalculationMethod.SignalToNoise.Spectroscopy(
-          exposures = exposures,
-          coadds = None,
-          exposureDuration = exposureDuration,
-          sourceFraction = 1.0,
-          ditherOffset = Angle.Angle0
-        ),
-        analysisMethod = observingMode.analysisMethod
-      ),
-      conditions = conditions,
-      telescope = ItcTelescopeDetails(
-        wfs = ItcWavefrontSensor.OIWFS
-      ),
-      instrument = ItcInstrumentDetails.fromObservingMode(observingMode)
-    )
 
 }
