+++ modules/itc/src/main/scala/lucuma/itc/search/ItcObservationDetails.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.itc
+package lucuma.itc.search
 
 import io.circe._
 import io.circe.generic.semiauto._
